prevent delete tech if it has rates

+++ server/src/application/handlers/technologies/delete.technology.handler.ts
@@ -1,6 +1,8 @@
 import TechnologyRepository from "../../../infrastructure/repositories/technology.repository";
+import RateRepository from "../../../infrastructure/repositories/rate.repository";
 import DeleteTechnologyCommand from "../../commands/technologies/delete.technology.command";
 import TechnologyNotFoundError from "../../customErrors/technologies/technology.not.found.error";
+import TechnologyHasRatesError from "../../customErrors/technologies/technology.has.rates.error";
 
 class DeleteTechnologyHandler{
     async execute(command: DeleteTechnologyCommand){
@@ -8,6 +10,10 @@ class DeleteTechnologyHandler{
 
         if (!technology) {
             throw new TechnologyNotFoundError();
+        }
+
+        if (await RateRepository.technologyHasRates(technology)) {
+            throw new TechnologyHasRatesError();
         }
 
         await TechnologyRepository.deleteById(technology.getId());
